requirements.txt
discord.py==0.16.12
requests==2.17.1
boto3==1.9.85